refactored sequence for mysql database
