+++ database-migration/src/main/resources/db/migration/V1_0_0_0__CREATE_TABLE_JIRA_CONNECTION.sql
@@ -1,7 +1,6 @@
 CREATE TABLE JIRA_CONNECTION_DATA(
-  ID INT NOT NULL,
+  ID INT NOT NULL AUTO_INCREMENT,
   URL VARCHAR(255) NOT NULL,
   MODIFIED_AT DATE NOT NULL,
   PRIMARY KEY (id)
 );
-CREATE SEQUENCE jira_connection_seq;
